// apps/tiktok-transformer/src/transformer/transformer.interface.ts
export interface IOrderDetailsRaw {
    buyer_email: string;
    buyer_message: string;
    cancel_order_sla_time: number;
    collection_due_time: number;
    commerce_platform: string;
    create_time: number;
    delivery_option_id: string;
    delivery_option_name: string;
    delivery_type: string;
    fulfillment_type: string;
    has_updated_recipient_address: boolean;
    id: string;
    is_cod: boolean;
    is_on_hold_order: boolean;
    is_replacement_order: boolean;
    is_sample_order: boolean;
    line_items: {
        currency: string;
        display_status: string;
        id: string;
        is_gift: boolean;
        original_price: string;
        package_id: string;
        package_status: string;
        platform_discount: string;
        product_id: string;
        product_name: string;
        rts_time: number;
        sale_price: string;
        seller_discount: string;
        seller_sku: string;
        shipping_provider_id: string;
        shipping_provider_name: string;
        sku_id: string;
        sku_image: string;
        sku_name: string;
        sku_type: string;
        tracking_number: string;
    }[];
    packages: {
        id: string;
    }[];
    payment: {
        currency: string;
        original_shipping_fee: string;
        original_total_product_price: string;
        platform_discount: string;
        seller_discount: string;
        shipping_fee: string;
        shipping_fee_cofunded_discount?: string;
        shipping_fee_platform_discount?: string;
        shipping_fee_seller_discount?: string;
        sub_total: string;
        tax: string;
        total_amount: string;
        handling_fee?: string;
        product_tax?: string;
        shipping_fee_tax?: string;
    };
    payment_method_name: string;
    recipient_address: {
        address_detail: string;
        address_line1: string;
        address_line2: string;
        address_line3: string;
        address_line4: string;
        district_info: {
            address_level: string;
            address_level_name: string;
            address_name: string;
        }[];
        first_name: string;
        first_name_local_script: string;
        full_address: string;
        last_name: string;
        last_name_local_script: string;
        name: string;
        phone_number: string;
        postal_code: string;
        region_code: string;
    };
    rts_sla_time: number;
    rts_time: number;
    shipping_due_time: number;
    shipping_provider: string;
    shipping_provider_id: string;
    shipping_type: string;
    status: string;
    tracking_number: string;
    tts_sla_time: number;
    update_time: number;
    user_id: string;
    warehouse_id: string;
}

export interface ITransformedOrderDetails {
    account: string;
    platform: string;
    shop_id: string;
    platform_order_id: string;
    api_version: string;
    buyer_id: number;
    billing_address: {
        address1: string;
        address2: string;
        address3: string;
        address4: string;
        city: string;
        postal_code: string;
        country: string;
        first_name: string;
        phone1: string;
    };
    buyer_coin: number;
    buyer_paid_shipping_fee: number;
    created_at: string;
    created_at_timestamp: number;
    credit_card_promotion: number;
    deleted: number;
    deleted_at: string;
    deleted_by: string;
    is_shop_active: boolean;
    items_count: number;
    net_buyer_price: number;
    order_number: string;
    payment_method: string;
    platform_buyer_username: string;
    platform_created_at: string;
    platform_created_at_timestamp: number;
    platform_updated_at: string;
    platform_updated_at_timetamp: number;
    price: number;
    promised_shipping_times: string;
    shipping_address: {
        address1: string;
        address2: string;
        address3: string;
        address4: string;
        city: string;
        postal_code: string;
        country: string;
        first_name: string;
        phone1: string;
    };
    shipping_fee: number;
    shipping_fee_discount_platform: number;
    shipping_fee_discount_seller: number;
    statuses: {
        items: any;
    };
    updated_at_timestamp: number;
    voucher: number;
    voucher_platform: number;
    voucher_seller: number;
}